+++ lib/src/controls/utils/hover_button.dart
@@ -7,22 +7,7 @@ typedef ButtonStateWidgetBuilder = Widget Function(
   Set<ButtonStates> state,
 );
 
-/// Base widget for any widget that requires input. It
-/// provides a [builder] callback to build the child with
-/// the current input state: none, hovering, pressing or
-/// focused.
-///
-/// It's used by the following widgets:
-/// - [Button]
-/// - [Checkbox]
-/// - [ComboBox]
-/// - [DatePicker]
-/// - [IconButton]
-/// - [RadioButton]
-/// - [TabView]'s [Tab]
-/// - [TappableListTile]
-/// - [TimePicker]
-/// - [ToggleSwitch]
+/// Base widget for any widget that requires input.
 class HoverButton extends StatefulWidget {
   /// Creates a hover button.
   const HoverButton({
@@ -48,6 +33,8 @@ class HoverButton extends StatefulWidget {
     this.customActions,
     this.shortcuts,
     this.focusEnabled = true,
+    this.forceEnabled = false,
+    this.hitTestBehavior = HitTestBehavior.opaque,
   }) : super(key: key);
 
   /// {@template fluent_ui.controls.inputs.HoverButton.mouseCursor}
@@ -58,6 +45,7 @@ class HoverButton extends StatefulWidget {
   /// cursor to the next region behind it in hit-test order.
   /// {@endtemplate}
   final MouseCursor? cursor;
+
   final VoidCallback? onLongPress;
   final VoidCallback? onLongPressStart;
   final VoidCallback? onLongPressEnd;
@@ -101,9 +89,10 @@ class HoverButton extends StatefulWidget {
   /// Whether actions are enabled
   ///
   /// Default actions:
-  ///  * Execute [onPressed] with Enter
+  ///  * Execute [onPressed] with Enter and Space
   ///
   /// See also:
+  ///
   ///  * [customActions], which lets you execute custom actions
   final bool actionsEnabled;
 
@@ -119,12 +108,23 @@ class HoverButton extends StatefulWidget {
 
   /// Whether the focusing is enabled.
   ///
-  /// If false, actions and shortcurts will not work, regardless of what is
+  /// If `false`, actions and shortcurts will not work, regardless of what is
   /// set on [actionsEnabled].
   final bool focusEnabled;
 
+  /// Whether the hover button should be always enabled.
+  ///
+  /// If `true`, the button will be considered active even if [onPressed] is not
+  /// provided
+  final bool forceEnabled;
+
+  /// How this gesture detector should behave during hit testing.
+  ///
+  /// This defaults to [HitTestBehavior.opaque]
+  final HitTestBehavior hitTestBehavior;
+
   @override
-  _HoverButtonState createState() => _HoverButtonState();
+  State<HoverButton> createState() => _HoverButtonState();
 }
 
 class _HoverButtonState extends State<HoverButton> {
@@ -190,16 +190,21 @@ class _HoverButtonState extends State<HoverButton> {
   bool _shouldShowFocus = false;
 
   bool get enabled =>
+      widget.forceEnabled ||
       widget.onPressed != null ||
       widget.onTapUp != null ||
       widget.onTapDown != null ||
       widget.onTapDown != null ||
       widget.onLongPress != null ||
       widget.onLongPressStart != null ||
-      widget.onLongPressEnd != null;
+      widget.onLongPressEnd != null ||
+      widget.onHorizontalDragStart != null ||
+      widget.onHorizontalDragUpdate != null ||
+      widget.onHorizontalDragEnd != null;
 
   Set<ButtonStates> get states {
     if (!enabled) return {ButtonStates.disabled};
+
     return {
       if (_pressing) ButtonStates.pressing,
       if (_hovering) ButtonStates.hovering,
@@ -210,7 +215,7 @@ class _HoverButtonState extends State<HoverButton> {
   @override
   Widget build(BuildContext context) {
     Widget w = GestureDetector(
-      behavior: HitTestBehavior.opaque,
+      behavior: widget.hitTestBehavior,
       onTap: enabled ? widget.onPressed : null,
       onTapDown: (_) {
         if (!enabled) return;